requirements.txt
feedparser
beautifulsoup4
python-dotenv
fpdf
google-api-python-client
google-auth
